refactor: Convert JavaScript files to TypeScript #3

+++ editor/src/Screens/HomeScreen/RightComponent/index.tsx
@@ -4,37 +4,53 @@ import { PlaygroundContext } from '../../../Providers/PlaygroundProvider'
 import { modalConstants, ModalContext } from '../../../Providers/ModalProvider'
 import { useNavigate } from 'react-router-dom'
 
-const Folder = ({ folderTitle, cards, folderId }) => {
-  const { deleteFolder, deleteFile } = useContext(PlaygroundContext)
-  const { openModal, setModalPayload } = useContext(ModalContext)
-  const navigate = useNavigate()
-  const onDeleteFolder = () => {
-    deleteFolder(folderId)
-  }
+type FileType = {
+  id: string;
+  title: string;
+};
 
+type FolderProps = {
+  folderTitle: string;
+  cards: FileType[];
+  folderId: string;
+};
+
+const Folder: React.FC<FolderProps> = ({ folderTitle, cards, folderId }) => {
+  const playgroundContext = useContext(PlaygroundContext)
+  const deleteFolder = playgroundContext?.deleteFolder
+  const deleteFile = playgroundContext?.deleteFile
+  const modalContext = useContext(ModalContext)
+  const openModal = modalContext?.openModal
+  const setModalPayload = modalContext?.setModalPayload
+  const navigate = useNavigate()
   const openCreateCardModal = () => {
-    setModalPayload(folderId)
-    openModal(modalConstants.CREATE_CARD)
+    setModalPayload && setModalPayload(folderId)
+    openModal && openModal(modalConstants.CREATE_CARD)
+  }
+  const onDeleteFolder = () => {
+    if (deleteFolder) {
+      deleteFolder(folderId)
+    }
   }
   return (
     <div className="folder-container">
       <div className="folder-header">
         <div className="folder-header-item">
-          <span class="material-symbols-outlined">folder_code</span>
+          <span className="material-symbols-outlined">folder_code</span>
           <span>{folderTitle}</span>
         </div>
         <div className="folder-header-item">
           <button>
-            <span class="material-symbols-outlined" onClick={onDeleteFolder}>
+            <span className="material-symbols-outlined" onClick={onDeleteFolder}>
               delete
             </span>
           </button>
 
           <button onClick={openCreateCardModal}>
-            <span class="material-symbols-outlined">add</span>
+            <span className="material-symbols-outlined">add</span>
           </button>
           <button>
-            <span class="material-symbols-outlined">
+            <span className="material-symbols-outlined">
               arrow_drop_down_circle
             </span>
           </button>
@@ -43,21 +59,23 @@ const Folder = ({ folderTitle, cards, folderId }) => {
       <div className="card-container">
         {cards?.map((file, index) => {
           const onDeleteFile = () => {
-            deleteFile(folderId, file.id)
+            if (deleteFile) {
+              deleteFile(folderId, file.id)
+            }
           }
           const navigateToPlayground = () => {
             navigate(`/playground/${folderId}/${file.id}`)
           }
           return (
             <div className="card" key={index} onClick={navigateToPlayground}>
-              <img src="code_blocks.png" />
+              <img src="/code_blocks.png" />
               <div className="card-title">
                 <span>{file?.title}</span>
               </div>
               <div className="card-element">
                 <button>
                   <span
-                    class="material-symbols-outlined"
+                    className="material-symbols-outlined"
                     onClick={onDeleteFile}
                   >
                     delete
@@ -73,8 +91,9 @@ const Folder = ({ folderTitle, cards, folderId }) => {
 }
 
 export const RightComponent = () => {
-  const { folders } = useContext(PlaygroundContext)
-  console.log(folders)
+  const playgroundContext = useContext(PlaygroundContext);
+  const folders = playgroundContext?.folders;
+  console.log(folders);
   return (
     <div className="right-container">
       <div className="header">
@@ -86,7 +105,6 @@ export const RightComponent = () => {
             folderTitle={folder?.title}
             cards={folder?.files}
             folderId={folder?.id}
-            fileId={folder?.files.id}
             key={index}
           />
         )
